Issue #386: DataCube.aggregate_spatial should return VectorCube
